Finalisation verifStockVitrine
Fonction finie et testée (le test est dans main en commentaire)
Produit avec constructeur et getter/setter

+++ src/Produit.java
@@ -48,7 +48,7 @@ public class Produit {
     }
 
 
-    public void setPrix(int prix)
+    public void setPrix(float prix)
     {
         this.prix=prix;
     }
